Crate gcp_bigquery_client

An ergonomic async client library for GCP BigQuery.

Support for dataset, table, streaming API and query (see status section for an exhaustive list of supported API endpoints)

Support Service Account Key authentication (other OAuth flows will be added later)

Create tables and rows via builder patterns

Persist complex Rust structs in structured BigQuery tables

Async API

Other OAuth flows will be added later.

For a detailed tutorial on the different ways to use GCP BigQuery Client please check out the GitHub repository.

Modules

auth

Helpers to manage GCP authentication.

client_builder

dataset

Manage BigQuery dataset.

error

List of the BigQuery errors supported by this crate.

job

Manage BigQuery jobs.

model

All the object definitions used by the BigQuery REST API.

model_api

Manage BigQuery models.

project

There is no persistent data associated with this resource.

routine

Manage BigQuery user-defined function or a stored procedure.

table

Manage BigQuery table

tabledata

Manage BigQuery streaming API.

Structs

Client

An asynchronous BigQuery client.

Functions

env_vars