Struct gcp_bigquery_client::job::JobApi

pub struct JobApi { /* private fields */ }

A job API handler.

Implementations

impl JobApi

pub async fn query(
&self,
project_id: &str,
query_request: QueryRequest
) -> Result<ResultSet, BQError>

Runs a BigQuery SQL query synchronously and returns query results if the query completes within a specified timeout.

Arguments

project_id - Project ID of the query request.

query_request - The request body contains an instance of QueryRequest.

pub fn query_all<'a>(
&'a self,
project_id: &'a str,
query: JobConfigurationQuery,
page_size: Option<i32>
) -> impl Stream<Item = Result<Vec<TableRow>, BQError>> + 'a

Runs a BigQuery SQL query, paginating through all the results synchronously.

Arguments

project_id- Project ID of the query request.

query - The initial query configuration that is submitted when the job is inserted.

page_size - The size of each page fetched. By default, this is set to None, and the limit is 10 MB of rows instead.

pub fn query_all_with_location<'a>(
&'a self,
project_id: &'a str,
location: &'a str,
query: JobConfigurationQuery,
page_size: Option<i32>
) -> impl Stream<Item = Result<Vec<TableRow>, BQError>> + 'a

Runs a BigQuery SQL query, paginating through all the results synchronously. Use this function when location of the job differs from the default value (US)

Arguments

project_id- Project ID of the query request.

location - Geographic location of the job.

query - The initial query configuration that is submitted when the job is inserted.

page_size - The size of each page fetched. By default, this is set to None, and the limit is 10 MB of rows instead.

pub fn query_all_with_job_reference<'a>(
&'a self,
project_id: &'a str,
job_reference: JobReference,
query: JobConfigurationQuery,
page_size: Option<i32>
) -> impl Stream<Item = Result<Vec<TableRow>, BQError>> + 'a

Runs a BigQuery SQL query, paginating through all the results synchronously. Use this function when you need to have your job with non-default location, project_id & job_id values

Arguments

project_id- Project ID of the query request.

job_reference - The initital job reference configuration that is submitted when the job is inserted

query - The initial query configuration that is submitted when the job is inserted.

page_size - The size of each page fetched. By default, this is set to None, and the limit is 10 MB of rows instead.

pub async fn insert(&self, project_id: &str, job: Job) -> Result<Job, BQError>

Starts a new asynchronous job.

Arguments

project_id - Project ID of project that will be billed for the job.

job - The request body contains an instance of Job.

pub async fn list(&self, project_id: &str) -> Result<JobList, BQError>

Lists all jobs that you started in the specified project. Job information is available for a six month period after creation. The job list is sorted in reverse chronological order, by job creation time. Requires the Can View project role, or the Is Owner project role if you set the allUsers property.

Arguments

project_id - Project ID of the jobs to list.

pub async fn get_query_results(
&self,
project_id: &str,
job_id: &str,
parameters: GetQueryResultsParameters
) -> Result<GetQueryResultsResponse, BQError>

RPC to get the results of a query job.

Arguments

project_id - Project ID of the query request.

job_id - Job ID of the query job.

parameters - The query parameters for jobs.getQueryResults.

pub async fn get_job(
&self,
project_id: &str,
job_id: &str,
location: Option<&str>
) -> Result<Job, BQError>

Returns information about a specific job. Job information is available for a six month period after creation. Requires that you’re the person who ran the job, or have the Is Owner project role.

Arguments

project_id - Project ID of the requested job.

job_id - Job ID of the requested job.

location - The geographic location of the job. Required except for US and EU. See details at https://cloud.google.com/bigquery/docs/locations#specifying_your_location.

pub async fn cancel_job(
&self,
project_id: &str,
job_id: &str,
location: Option<&str>
) -> Result<JobCancelResponse, BQError>

Requests that a job be cancelled. This call will return immediately, and the client will need to poll for the job status to see if the cancel completed successfully. Cancelled jobs may still incur costs.

Arguments

project_id - Project ID of the job to cancel.

job_id - Job ID of the job to cancel.

location - The geographic location of the job. Required except for US and EU. See details at https://cloud.google.com/bigquery/docs/locations#specifying_your_location.