Module gcp_bigquery_client::model

All the object definitions used by the BigQuery REST API.

Modules

aggregate_classification_metrics

Aggregate metrics for classification/classifier models. For multi-class models, the metrics are either macro-averaged or micro-averaged. When macro-averaged, the metrics are calculated for each label and then an unweighted average is taken of those values. When micro-averaged, the metric is calculated globally by counting the total number of correctly predicted rows.

argument

Input/output argument of a function or a stored procedure.

arima_coefficients

Arima coefficients.

arima_fitting_metrics

ARIMA model fitting metrics.

arima_forecasting_metrics

Model evaluation metrics for ARIMA forecasting models.

arima_model_info

Arima model information.

arima_order

Arima order, can be used for both non-seasonal and seasonal parts.

arima_result

(Auto-)arima fitting result. Wrap everything in ArimaResult for easier refactoring if we want to use model-specific iteration results.

arima_single_model_forecasting_metrics

Model evaluation metrics for a single ARIMA forecasting model.

audit_config

audit_log_config

bigquery_model_training

bigtable_column

bigtable_column_family

bigtable_options

binary_classification_metrics

Evaluation metrics for binary classification/classifier models.

binary_confusion_matrix

Confusion matrix for binary classification models.

binding

bqml_iteration_result

bqml_training_run

bqml_training_run_training_options

categorical_value

Representative value of a categorical feature.

category_count

Represents the count of a single category within the cluster.

cluster

Message containing the information about one cluster.

cluster_info

Information about a single cluster for clustering model.

clustering

clustering_metrics

Evaluation metrics for clustering models.

confusion_matrix

Confusion matrix for multi-class classification models.

connection_property

csv_options

data_format_options

data_split_result

Data split result. This contains references to the training and evaluation data tables that were used to train the model.

dataset

dataset_reference

datasets

destination_table_properties

dimensionality_reduction_metrics

Model evaluation metrics for dimensionality reduction models.

encryption_configuration

entry

A single entry in the confusion matrix.

error_proto

evaluation_metrics

Evaluation metrics of a model. These are either computed on all training data or just the eval data based on whether eval data was used during training. These are not present for imported models.

explain_query_stage

explain_query_step

explanation

Explanation for a single feature.

expr

external_data_configuration

feature_value

Representative value of a single feature within the cluster.

field_type

get_iam_policy_request

get_policy_options

get_query_results_parameters

get_query_results_response

get_service_account_response

global_explanation

Global explanations containing the top most important features after training.

google_sheets_options

hive_partitioning_options

information_schema

iteration_result

Information about a single iteration of the training run.

job

job_cancel_response

job_configuration

job_configuration_extract

job_configuration_load

job_configuration_query

job_configuration_table_copy

job_list

job_list_jobs

job_reference

job_statistics

job_statistics2

job_statistics3

job_statistics4

job_statistics_reservation_usage

job_status

list_models_response

list_routines_response

materialized_view_definition

model

model_definition

model_definition_model_options

model_reference

multi_class_classification_metrics

Evaluation metrics for multi-class classification/classifier models.

policy

principal_component_info

Principal component infos, used only for eigen decomposition based models, e.g., PCA. Ordered by explained_variance in the descending order.

project_list

project_reference

query_parameter

query_parameter_type

query_parameter_type_struct_types

query_parameter_value

query_request

query_response

query_timeline_sample

range_partitioning

range_partitioning_range

ranking_metrics

Evaluation metrics used by weighted-ALS models specified by feedback_type=implicit.

regression_metrics

Evaluation metrics for regression and explicit feedback type matrix factorization models.

routine

A user-defined function or a stored procedure.

routine_reference

row

A single row in the confusion matrix.

row_access_policy

Represents access on a subset of rows on the specified table, defined by its filter predicate. Access to the subset of rows is controlled by its IAM policy.

row_access_policy_reference

row_level_security_statistics

script_stack_frame

script_statistics

set_iam_policy_request

snapshot_definition

standard_sql_data_type

The type of a variable, e.g., a function argument. Examples: INT64: {type_kind=“INT64”} ARRAY: {type_kind=“ARRAY”, array_element_type=“STRING”} STRUCT>: {type_kind=“STRUCT”, struct_type={fields=[ {name=“x”, type={type_kind=“STRING”}}, {name=“y”, type={type_kind=“ARRAY”, array_element_type=“DATE”}} ]}}

standard_sql_field

A field or a column.

standard_sql_struct_type

streamingbuffer

table

table_cell

table_data_insert_all_request

table_data_insert_all_request_rows

table_data_insert_all_response

table_data_insert_all_response_insert_errors

table_field_schema

table_field_schema_categories

table_field_schema_policy

table_list

table_list_tables

table_list_view

table_reference

table_row

table_schema

test_iam_permissions_request

test_iam_permissions_response

time_partitioning

training_options

Options used in model training.

training_run

Information about a single training query run for the model.

transaction_info

user_defined_function_resource

view_definition