Struct gcp_bigquery_client::model::job_statistics2::JobStatistics2

pub struct JobStatistics2 {

    pub billing_tier: Option<i32>,
    pub cache_hit: Option<bool>,
    pub ddl_affected_row_access_policy_count: Option<String>,
    pub ddl_operation_performed: Option<String>,
    pub ddl_target_routine: Option<RoutineReference>,
    pub ddl_target_row_access_policy: Option<RowAccessPolicyReference>,
    pub ddl_target_table: Option<TableReference>,
    pub estimated_bytes_processed: Option<String>,
    pub model_training: Option<BigQueryModelTraining>,
    pub model_training_current_iteration: Option<i32>,
    pub model_training_expected_total_iteration: Option<String>,
    pub num_dml_affected_rows: Option<String>,
    pub query_plan: Option<Vec<ExplainQueryStage>>,
    pub referenced_routines: Option<Vec<RoutineReference>>,
    pub referenced_tables: Option<Vec<TableReference>>,
    pub reservation_usage: Option<Vec<JobStatisticsReservationUsage>>,
    pub schema: Option<TableSchema>,
    pub statement_type: Option<String>,
    pub timeline: Option<Vec<QueryTimelineSample>>,
    pub total_bytes_billed: Option<String>,
    pub total_bytes_processed: Option<String>,
    pub total_bytes_processed_accuracy: Option<String>,
    pub total_partitions_processed: Option<String>,
    pub total_slot_ms: Option<String>,
    pub undeclared_query_parameters: Option<Vec<QueryParameter>>,

}

Fields
billing_tier: Option<i32>

[Output-only] Billing tier for the job.

cache_hit: Option<bool>

[Output-only] Whether the query result was fetched from the query cache.

ddl_affected_row_access_policy_count: Option<String>

[Output-only] [Preview] The number of row access policies affected by a DDL statement. Present only for DROP ALL ROW ACCESS POLICIES queries.

ddl_operation_performed: Option<String>

The DDL operation performed, possibly dependent on the pre-existence of the DDL target. Possible values (new values might be added in the future): "CREATE": The query created the DDL target. "SKIP": No-op. Example cases: the query is CREATE TABLE IF NOT EXISTS while the table already exists, or the query is DROP TABLE IF EXISTS while the table does not exist. "REPLACE": The query replaced the DDL target. Example case: the query is CREATE OR REPLACE TABLE, and the table already exists. "DROP": The query deleted the DDL target.

ddl_target_routine: Option<RoutineReference>ddl_target_row_access_policy: Option<RowAccessPolicyReference>ddl_target_table: Option<TableReference>estimated_bytes_processed: Option<String>

[Output-only] The original estimate of bytes processed for the job.

model_training: Option<BigQueryModelTraining>model_training_current_iteration: Option<i32>

[Output-only, Beta] Deprecated; do not use.

model_training_expected_total_iteration: Option<String>

[Output-only, Beta] Deprecated; do not use.

num_dml_affected_rows: Option<String>

[Output-only] The number of rows affected by a DML statement. Present only for DML statements INSERT, UPDATE or DELETE.

query_plan: Option<Vec<ExplainQueryStage>>

[Output-only] Describes execution plan for the query.

referenced_routines: Option<Vec<RoutineReference>>

[Output-only] Referenced routines (persistent user-defined functions and stored procedures) for the job.

referenced_tables: Option<Vec<TableReference>>

[Output-only] Referenced tables for the job. Queries that reference more than 50 tables will not have a complete list.

reservation_usage: Option<Vec<JobStatisticsReservationUsage>>

[Output-only] Job resource usage breakdown by reservation.

schema: Option<TableSchema>statement_type: Option<String>

The type of query statement, if valid. Possible values (new values might be added in the future): "SELECT": SELECT query. "INSERT": INSERT query; see https://cloud.google.com/bigquery/docs/reference/standard-sql/data-manipulation-language. "UPDATE": UPDATE query; see https://cloud.google.com/bigquery/docs/reference/standard-sql/data-manipulation-language. "DELETE": DELETE query; see https://cloud.google.com/bigquery/docs/reference/standard-sql/data-manipulation-language. "MERGE": MERGE query; see https://cloud.google.com/bigquery/docs/reference/standard-sql/data-manipulation-language. "ALTER_TABLE": ALTER TABLE query. "ALTER_VIEW": ALTER VIEW query. "ASSERT": ASSERT condition AS ‘description’. "CREATE_FUNCTION": CREATE FUNCTION query. "CREATE_MODEL": CREATE [OR REPLACE] MODEL … AS SELECT … . "CREATE_PROCEDURE": CREATE PROCEDURE query. "CREATE_TABLE": CREATE [OR REPLACE] TABLE without AS SELECT. "CREATE_TABLE_AS_SELECT": CREATE [OR REPLACE] TABLE … AS SELECT … . "CREATE_VIEW": CREATE [OR REPLACE] VIEW … AS SELECT … . "DROP_FUNCTION" : DROP FUNCTION query. "DROP_PROCEDURE": DROP PROCEDURE query. "DROP_TABLE": DROP TABLE query. "DROP_VIEW": DROP VIEW query.

timeline: Option<Vec<QueryTimelineSample>>

[Output-only] [Beta] Describes a timeline of job execution.

total_bytes_billed: Option<String>

[Output-only] Total bytes billed for the job.

total_bytes_processed: Option<String>

[Output-only] Total bytes processed for the job.

total_bytes_processed_accuracy: Option<String>

[Output-only] For dry-run jobs, totalBytesProcessed is an estimate and this field specifies the accuracy of the estimate. Possible values can be: UNKNOWN: accuracy of the estimate is unknown. PRECISE: estimate is precise. LOWER_BOUND: estimate is lower bound of what the query would cost. UPPER_BOUND: estimate is upper bound of what the query would cost.

total_partitions_processed: Option<String>

[Output-only] Total number of partitions processed from all partitioned tables referenced in the job.

total_slot_ms: Option<String>

[Output-only] Slot-milliseconds for the job.

undeclared_query_parameters: Option<Vec<QueryParameter>>

Standard SQL only: list of undeclared query parameters detected during a dry run validation.