Struct gcp_bigquery_client::model::query_response::QueryResponse

pub struct QueryResponse {
    pub cache_hit: Option<bool>,
    pub errors: Option<Vec<ErrorProto>>,
    pub job_complete: Option<bool>,
    pub job_reference: Option<JobReference>,
    pub kind: Option<String>,
    pub num_dml_affected_rows: Option<String>,
    pub page_token: Option<String>,
    pub rows: Option<Vec<TableRow>>,
    pub schema: Option<TableSchema>,
    pub total_bytes_processed: Option<String>,
    pub total_rows: Option<String>,
}

Fields
cache_hit: Option<bool>

Whether the query result was fetched from the query cache.

errors: Option<Vec<ErrorProto>>

[Output-only] The first errors or warnings encountered during the running of the job. The final message includes the number of errors that caused the process to stop. Errors here do not necessarily mean that the job has completed or was unsuccessful.

job_complete: Option<bool>

Whether the query has completed or not. If rows or totalRows are present, this will always be true. If this is false, totalRows will not be available.

job_reference: Option<JobReference>kind: Option<String>

The resource type.

num_dml_affected_rows: Option<String>

[Output-only] The number of rows affected by a DML statement. Present only for DML statements INSERT, UPDATE or DELETE.

page_token: Option<String>

A token used for paging results.

rows: Option<Vec<TableRow>>

An object with as many results as can be contained within the maximum permitted reply size. To get any additional rows, you can call GetQueryResults and specify the jobReference returned above.

schema: Option<TableSchema>total_bytes_processed: Option<String>

The total number of bytes processed for this query. If this query was a dry run, this is the number of bytes that would be processed if the query were run.

total_rows: Option<String>

The total number of rows in the complete query result set, which can be more than the number of rows in this single page of results.