Struct gcp_bigquery_client::model::query_response::ResultSet

pub struct ResultSet { /* private fields */ }

Set of rows in response to a SQL query

Implementations

impl ResultSet

pub fn new(query_response: QueryResponse) -> Self

pub fn query_response(&self) -> &QueryResponse

pub fn next_row(&mut self) -> bool

Moves the cursor froward one row from its current position. A ResultSet cursor is initially positioned before the first row; the first call to the method next makes the first row the current row; the second call makes the second row the current row, and so on.

pub fn row_count(&self) -> usize

Total number of rows in this result set.

pub fn column_names(&self) -> Vec<String>

List of column names for this result set.

pub fn column_index(&self, column_name: &str) -> Option<&usize>

Returns the index for a column name.

pub fn get_i64(&self, col_index: usize) -> Result<Option<i64>, BQError>

pub fn get_i64_by_name(&self, col_name: &str) -> Result<Option<i64>, BQError>

pub fn get_f64(&self, col_index: usize) -> Result<Option<f64>, BQError>

pub fn get_f64_by_name(&self, col_name: &str) -> Result<Option<f64>, BQError>

pub fn get_bool(&self, col_index: usize) -> Result<Option<bool>, BQError>

pub fn get_bool_by_name(&self, col_name: &str) -> Result<Option<bool>, BQError>

pub fn get_string(&self, col_index: usize) -> Result<Option<String>, BQError>

pub fn get_string_by_name(
&self,
col_name: &str
) -> Result<Option<String>, BQError>

pub fn get_json_value(&self, col_index: usize) -> Result<Option<Value>, BQError>

pub fn get_json_value_by_name(
&self,
col_name: &str
) -> Result<Option<Value>, BQError>