Struct gcp_bigquery_client::model::table::Table

pub struct Table {

    pub clustering: Option<Clustering>,
    pub creation_time: Option<String>,
    pub description: Option<String>,
    pub encryption_configuration: Option<EncryptionConfiguration>,
    pub etag: Option<String>,
    pub expiration_time: Option<String>,
    pub external_data_configuration: Option<ExternalDataConfiguration>,
    pub friendly_name: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub last_modified_time: Option<String>,
    pub location: Option<String>,
    pub materialized_view: Option<MaterializedViewDefinition>,
    pub model: Option<ModelDefinition>,
    pub num_bytes: Option<String>,
    pub num_long_term_bytes: Option<String>,
    pub num_physical_bytes: Option<String>,
    pub num_rows: Option<String>,
    pub range_partitioning: Option<RangePartitioning>,
    pub require_partition_filter: Option<bool>,
    pub schema: TableSchema,
    pub self_link: Option<String>,
    pub snapshot_definition: Option<SnapshotDefinition>,
    pub streaming_buffer: Option<Streamingbuffer>,
    pub table_reference: TableReference,
    pub time_partitioning: Option<TimePartitioning>,
    pub type: Option<String>,
    pub view: Option<ViewDefinition>,

}

Fields
clustering: Option<Clustering>creation_time: Option<String>

[Output-only] The time when this table was created, in milliseconds since the epoch.

description: Option<String>

[Optional] A user-friendly description of this table.

encryption_configuration: Option<EncryptionConfiguration>etag: Option<String>

[Output-only] A hash of the table metadata. Used to ensure there were no concurrent modifications to the resource when attempting an update. Not guaranteed to change when the table contents or the fields numRows, numBytes, numLongTermBytes or lastModifiedTime change.

expiration_time: Option<String>

[Optional] The time when this table expires, in milliseconds since the epoch. If not present, the table will persist indefinitely. Expired tables will be deleted and their storage reclaimed. The defaultTableExpirationMs property of the encapsulating dataset can be used to set a default expirationTime on newly created tables.

external_data_configuration: Option<ExternalDataConfiguration>friendly_name: Option<String>

[Optional] A descriptive name for this table.

id: Option<String>

[Output-only] An opaque ID uniquely identifying the table.

kind: Option<String>

[Output-only] The type of the resource.

labels: Option<HashMap<String, String>>

The labels associated with this table. You can use these to organize and group your tables. Label keys and values can be no longer than 63 characters, can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. Label values are optional. Label keys must start with a letter and each label in the list must have a different key.

last_modified_time: Option<String>

[Output-only] The time when this table was last modified, in milliseconds since the epoch.

location: Option<String>

[Output-only] The geographic location where the table resides. This value is inherited from the dataset.

materialized_view: Option<MaterializedViewDefinition>model: Option<ModelDefinition>num_bytes: Option<String>

[Output-only] The size of this table in bytes, excluding any data in the streaming buffer.

num_long_term_bytes: Option<String>

[Output-only] The number of bytes in the table that are considered "long-term storage".

num_physical_bytes: Option<String>

[Output-only] [TrustedTester] The physical size of this table in bytes, excluding any data in the streaming buffer. This includes compression and storage used for time travel.

num_rows: Option<String>

[Output-only] The number of rows of data in this table, excluding any data in the streaming buffer.

range_partitioning: Option<RangePartitioning>require_partition_filter: Option<bool>

[Optional] If set to true, queries over this table require a partition filter that can be used for partition elimination to be specified.

schema: TableSchemaself_link: Option<String>

[Output-only] A URL that can be used to access this resource again.

snapshot_definition: Option<SnapshotDefinition>streaming_buffer: Option<Streamingbuffer>table_reference: TableReferencetime_partitioning: Option<TimePartitioning>type: Option<String>

[Output-only] Describes the table type. The following values are supported: TABLE: A normal BigQuery table. VIEW: A virtual table defined by a SQL query. SNAPSHOT: An immutable, read-only table that is a copy of another table. [TrustedTester] MATERIALIZED_VIEW: SQL query whose result is persisted. EXTERNAL: A table that references data stored in an external storage system, such as Google Cloud Storage. The default value is TABLE.

view: Option<ViewDefinition>
Implementations

impl Table

pub fn new(
project_id: &str,
dataset_id: &str,
table_id: &str,
schema: TableSchema
) -> Self

pub fn from_dataset(
dataset: &Dataset,
table_id: &str,
schema: TableSchema
) -> Self

Creates a table struct pre-initialized with the project id and dataset id coming from the dataset passed in parameter.

Arguments

dataset - The dataset where the table will be located

table_id - The table name

schema - The table schema definition

pub fn project_id(&self) -> &String

Returns the project id of table.

pub fn dataset_id(&self) -> &String

Returns the dataset id of table.

pub fn table_id(&self) -> &String

Returns the table id of table.

pub fn location(self, location: &str) -> Self

Sets the location of this table.

Arguments

location - The location of this table

pub fn friendly_name(self, friendly_name: &str) -> Self

Sets the friendly name of this table.

Arguments

friendly_name - The friendly name of this table

pub fn description(self, description: &str) -> Self

Sets the description of this table.

Arguments

description - The description of this table

pub fn time_partitioning(self, time_partitioning: TimePartitioning) -> Self

pub fn range_partitioning(self, range_partitioning: RangePartitioning) -> Self

pub fn clustering(self, clustering: Clustering) -> Self

pub fn require_partition_filter(self, require_partition_filter: bool) -> Self

pub fn expiration_time(self, expiration_time: SystemTime) -> Self

pub fn view(self, view: ViewDefinition) -> Self

pub fn materialized_view(
self,
materialized_view: MaterializedViewDefinition
) -> Self

pub fn external_data_configuration(
self,
external_data_configuration: ExternalDataConfiguration
) -> Self

pub fn encryption_configuration(
self,
encryption_configuration: EncryptionConfiguration
) -> Self

pub fn snapshot_definition(self, snapshot_definition: SnapshotDefinition) -> Self

pub fn label(self, key: &str, value: &str) -> Self

Adds a label to this table.

Arguments

key - The label name.

value - The label value.

pub async fn delete(self, client: &Client) -> Result<(), BQError>