Enum gcp_bigquery_client::model::training_options::HolidayRegion

pub enum HolidayRegion {

    HolidayRegionUnspecified,
    Global,
    Na,
    Japac,
    Emea,
    Lac,
    Ae,
    Ar,
    At,
    Au,
    Be,
    Br,
    Ca,
    Ch,
    Cl,
    Cn,
    Co,
    C,
    Cz,
    De,
    Dk,
    Dz,
    Ec,
    Ee,
    Eg,
    E,
    Fi,
    Fr,
    Gb,
    Gr,
    Hk,
    Hu,
    Id,
    Ie,
    Il,
    In,
    Ir,
    It,
    Jp,
    Kr,
    Lv,
    Ma,
    Mx,
    My,
    Ng,
    Nl,
    No,
    Nz,
    Pe,
    Ph,
    Pk,
    Pl,
    Pt,
    Ro,
    R,
    Ru,
    Sa,
    Se,
    Sg,
    Si,
    Sk,
    Th,
    Tr,
    Tw,
    Ua,
    U,
    Ve,
    Vn,
    Za,

}

The geographical region based on which the holidays are considered in time series modeling. If a valid value is specified, then holiday effects modeling is enabled.

Variants

HolidayRegionUnspecified

Holiday region unspecified.

Global

Global.

Na

North America.

Japac

Japan and Asia Pacific: Korea, Greater China, India, Australia, and New Zealand.

Emea

Europe, the Middle East and Africa.

Lac

Latin America and the Caribbean.

Ae

United Arab Emirates

Ar

Argentina

At

Austria

Au

Australia

Be

Belgium

Br

Brazil

Ca

Canada

Ch

Switzerland

Cl

Chile

Cn

China

Co

Colombia

C

Czechoslovakia

Cz

Czech Republic

De

Germany

Dk

Denmark

Dz

Algeria

Ec

Ecuador

Ee

Estonia

Eg

Egypt

E

Spain

Fi

Finland

Fr

France

Gb

Great Britain (United Kingdom)

Gr

Greece

Hk

Hong Kong

Hu

Hungary

Id

Indonesia

Ie

Ireland

Il

Israel

In

India

Ir

Iran

It

Italy

Jp

Japan

Kr

Korea (South)

Lv

Latvia

Ma

Morocco

Mx

Mexico

My

Malaysia

Ng

Nigeria

Nl

Netherlands

No

Norway

Nz

New Zealand

Pe

Peru

Ph

Philippines

Pk

Pakistan

Pl

Poland

Pt

Portugal

Ro

Romania

R

Serbia

Ru

Russian Federation

Sa

Saudi Arabia

Se

Sweden

Sg

Singapore

Si

Slovenia

Sk

Slovakia

Th

Thailand

Tr

Turkey

Tw

Taiwan

Ua

Ukraine

U

United States

Ve

Venezuela

Vn

Viet Nam

Za

South Africa